Factor out component PaginatedTable
