+++ src/components/controls/PaginatedTable/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "PaginatedTable",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./PaginatedTable.js"
+}
